Updated the logout menu to be closed whenever the user will click outside the menu itself
