+++ src/components/layout/app-layout/components/header-link.tsx
@@ -1,0 +1,16 @@
+import { Link } from "react-router-dom";
+
+type Props = {
+  text: string;
+  location: string;
+}
+
+export function HeaderLink(props: Props) {
+  return (
+    <li>
+      <Link to={props.location} className={'block transition hover:text-primary md:px-4'}>
+        {props.text}
+      </Link>
+    </li>
+  )
+}
