+++ src/components/layout/app-layout/components/index.ts
@@ -1,0 +1,3 @@
+export * from './copy-icon'
+export * from './header-link'
+export * from './logout-menu'
